unifdef: use memcpy instead of strncpy
commit 38c7b224ce22c25fed04007839edf974bd13439d upstream.

New versions of gcc reasonably warn about the odd pattern of

	strncpy(p, q, strlen(q));

which really doesn't make sense: the strncpy() ends up being just a slow
and odd way to write memcpy() in this case.

There was a comment about _why_ the code used strncpy - to avoid the
terminating NUL byte, but memcpy does the same and avoids the warning.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

[sw0312.kim: cherry-pick stable linux-3.18.y commit c6c893b2dabf for gcc 9 build]
Signed-off-by: Seung-Woo Kim <sw0312.kim@samsung.com>

Change-Id: I220d1195b95a76841c6efce4f16455640191406a