motor: zh915: Update level value according to new motor interface
Priot to that 'u.rumble.strong_magnitude' contains the value as following:
- In case of intensity is strong : 65536
- In case of insensity is weak, 32767

But, the new motor interface uses 'u.rumble.strong_magnitude' as the
'intensity_level' which is range from 1 to 5. In order to keep
the compatibility, calculate the level in the kernel according to
the 'intensity_level' from user process (feedbackd).

In result, the 'level' would be initialized as following to keep
the compatibility.
- If level is 1 (strong), the calculated level value is 65536.
- If level is 2 (weak), the calculated level value is 32767.

Change-Id: Ia8f2fc730432ce9a6c589d571bd9c629627b9316
Signed-off-by: Chanwoo Choi <cw00.choi@samsung.com>